requirements.txt
pytest>=8.3.5
pytest-xdist>=3.6.1
pytest-html>=4.1.1
selenium>=4.29.0
openpyxl>=3.1.5
PyYAML>=6.0.2
allure-pytest>=2.13.5

